Move files around.
