+++ assets/src/edit-story/components/canvas/singleSelectionMoveable/useResize.js
